refactor(example): split React chat
