+++ examples/react-chat/package.json
@@ -3,27 +3,10 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "agent": "bun run build && bun run ../../apps/cli/src/main.ts dev --no-open --ui dist",
-    "build": "vite build",
-    "dev": "vite",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@base-ui/react": "^1.7.0",
-    "@clavia/tardigrade-client": "workspace:*",
-    "@clavia/tardigrade-ui": "workspace:*",
-    "@phosphor-icons/react": "^2.1.10",
-    "@tanstack/react-query": "^5.102.8",
-    "lucide-react": "^1.40.0",
-    "react": "^19.2.8",
-    "react-dom": "^19.2.8",
-    "tardie": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/react": "^19.2.18",
-    "@types/react-dom": "^19.2.7",
-    "@vitejs/plugin-react": "^6.1.1",
-    "typescript": "^7.0.2",
-    "vite": "^8.2.2"
+    "build": "bun run --cwd web build",
+    "dev": "bun run --filter '@clavia/tardigrade-react-chat-*' --parallel dev",
+    "dev:server": "bun run --cwd server dev",
+    "dev:web": "bun run --cwd web dev",
+    "typecheck": "bun run --cwd server typecheck && bun run --cwd web typecheck"
   }
 }
