+++ examples/react-chat/web/src/main.tsx
@@ -1,0 +1,16 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import { StrictMode } from "react"
+import { createRoot } from "react-dom/client"
+
+import { Chat } from "./components/Chat"
+import "./styles.css"
+
+const queryClient = new QueryClient()
+
+createRoot(document.getElementById("root")!).render(
+  <StrictMode>
+    <QueryClientProvider client={queryClient}>
+      <Chat />
+    </QueryClientProvider>
+  </StrictMode>
+)
