+++ examples/react-chat/web/vite.config.ts
