--- examples/react-chat/wrangler.jsonc
@@ -1,22 +1,0 @@
-{
-  "vars": {
-    "TARDIGRADE_CONFIG": {
-      "models": {
-        "allow": "*",
-        "providers": {
-          "openrouter": {
-            "baseUrl": "https://openrouter.ai/api/v1",
-            "protocol": "openai-chat-completions",
-            "env": [
-              "OPENROUTER_API_KEY"
-            ]
-          }
-        },
-        "default": {
-          "provider": "openrouter",
-          "model_id": "openai/gpt-5-mini"
-        }
-      }
-    }
-  }
-}
