+++ package.json
@@ -10,7 +10,7 @@
     "platform/*",
     "apps/*",
     "e2e",
-    "examples/react-chat"
+    "examples/react-chat/*"
   ],
   "scripts": {
     "tdg": "bun run apps/cli/src/main.ts",
